requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.9
colorama==0.4.4
decorator==4.4.2
ffmpeg==1.4
idna==3.3
imageio==2.13.4
imageio-ffmpeg==0.4.5
moviepy==1.0.3
numpy==1.21.5
Pillow==8.4.0
proglog==0.1.9
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
tqdm==4.62.3
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
